Use a deque.
We need to extend the child nodes from the left and take nodes from the
right. Currently it worked like a stack, which would be a DFS. That way
I'm not extending the search space first before I even checked all child
nodes. This is important as I'd like to continue the search.